파티방 카드에 대해서 모듈화 적용 완료
